Rename definition file
